+++ earcrate/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "0.8.28"
+__version__ = "0.8.29"
 from earcrate.cli import main